requirements.txt
dotenv
azure-identity
agent-framework
chainlit